Modifiche

+++ laureando.cpp
@@ -2,7 +2,7 @@
 
 unsigned short Laureando::bonusVoto = 200;
 
-Laureando::Laureando(std::string n, unsigned short e, corsoLaurea c, bool r, unsigned short m, unsigned short v)
+Laureando::Laureando(std::string n, unsigned short e, corsoLaurea c, bool r, double m, unsigned short v)
     : Persona(n,e), Studente(n,e,c,r,m), votoBase(v) {}
 
 unsigned short Laureando::getVotoBase() const
